adk_hackathon_streamlit_GeminiVision/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
google-cloud-aiplatform>=1.50.0
google-cloud-secret-manager>=2.0.0
google-cloud-bigquery>=3.0.0
anthropic>=0.25.0
praw>=7.7.0
streamlit>=1.35.0
fastapi>=0.110.0
uvicorn>=0.28.0
pillow>=10.0.0
ffmpeg-python>=0.2.0
requests>=2.31.0
toml>=0.10.0
python-dotenv>=1.0.0
